requirements.txt
# Core dependencies
python-dotenv>=1.0.0
click>=8.0.0
python-box>=7.0.0
pydantic>=2.5.0

# Azure AI Foundry and services
azure-ai-ml>=1.12.0
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0
azure-cognitiveservices-language-textanalytics>=5.3.0
azure-search-documents>=11.4.0
azure-ai-documentintelligence>=1.0.0
azure-ai-contentsafety>=1.0.0
openai>=1.12.0

# Data processing and embeddings
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
chromadb>=0.4.0
faiss-cpu>=1.7.0

# ML and AI frameworks
transformers>=4.36.0
torch>=2.1.0
datasets>=2.16.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
jupyter>=1.0.0

# MLOps and monitoring
mlflow>=2.9.0
wandb>=0.16.0
azure-monitor-opentelemetry>=1.2.0
opentelemetry-api>=1.20.0

# Testing and quality
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
bandit>=1.7.0

# Web framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
aiofiles>=23.0.0

# Security and safety
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Development and utilities
pre-commit>=3.6.0
tqdm>=4.66.0
python-multipart>=0.0.6
jinja2>=3.1.0